+++ bin/run_plugin_test.sh
@@ -17,4 +17,4 @@ done
 root="$( cd -P "$( dirname "$source" )" && pwd )"
 
 cd "$root/.."
-"./check_pages" example/tests/runner_plugins.php --filter=$1 $@
+"./check_pages" run example/tests/runner_plugins.php --filter=$1 $@