+++ src/Event/RunnerEventInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace AKlump\CheckPages\Event;
+
+interface RunnerEventInterface {
+
+  public function getRunner(): \AKlump\CheckPages\Parts\Runner;
+}